Restructure folders for witness related code
